Validação forms products

+++ database/migrations/2025_04_05_160127_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->integer('quantity');
             $table->decimal('price', 8, 2);
             $table->unsignedBigInteger('category_id');
-            $table->decimal('promo')->nullable();
+            $table->integer('promo')->nullable();
             $table->timestamps();
 
             $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
